extracting composables from tags.ts
